# Use optimized binary search for non-trivial types in B+Tree